multi-app zipper
