+++ core/frame/handshake_frame_test.go
@@ -9,11 +9,12 @@ import (
 func TestHandshakeFrameEncode(t *testing.T) {
 	expectedName := "1234"
 	var expectedType byte = 0xD3
-	m := NewHandshakeFrame(expectedName, expectedType, 0x0, nil)
+	m := NewHandshakeFrame(expectedName, expectedType, "", 0x0, nil)
 	assert.Equal(t, []byte{
-		0x80 | byte(TagOfHandshakeFrame), 0x0e,
+		0x80 | byte(TagOfHandshakeFrame), 0x10,
 		byte(TagOfHandshakeName), 0x04, 0x31, 0x32, 0x33, 0x34,
 		byte(TagOfHandshakeType), 0x01, 0xD3,
+		byte(TagOfHandshakeAppID), 0x0,
 		byte(TagOfHandshakeAuthType), 0x01, 0x0,
 		byte(TagOfHandshakeAuthPayload), 0x0,
 	},
