+++ zipper.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"sync"
 
 	"github.com/yomorun/yomo/core"
 	"github.com/yomorun/yomo/pkg/config"
@@ -129,13 +128,6 @@ func (z *zipper) ConfigWorkflow(conf string) error {
 }
 
 func (z *zipper) configWorkflow(config *config.WorkflowConfig) error {
-	// for i, app := range config.Functions {
-	// 	if err := z.server.AddWorkflow(core.Workflow{Seq: i, Name: app.Name}); err != nil {
-	// 		return err
-	// 	}
-	// 	logger.Printf("%s[AddWorkflow] %d, %s", zipperLogPrefix, i, app.Name)
-	// }
-	// return nil
 	return z.server.ConfigRouter(newRouter(config))
 }
 
@@ -242,11 +234,6 @@ func (z *zipper) Close() error {
 func (z *zipper) Stats() int {
 	log.Printf("[%s] all sfn connected: %d", z.name, len(z.server.StatsFunctions()))
 	for k, _ := range z.server.StatsFunctions() {
-		// ids := make([]int64, 0)
-		// for _, c := range v {
-		// 	ids = append(ids, int64((*c).StreamID()))
-		// }
-		// log.Printf("[%s] -> k=%v, v.StreamID=%v", z.name, k, ids)
 		log.Printf("[%s] -> ConnID=%v", z.name, k)
 	}
 
@@ -258,73 +245,4 @@ func (z *zipper) Stats() int {
 	log.Printf("[%s] total DataFrames received: %d", z.name, z.server.StatsCounter())
 
 	return len(z.server.StatsFunctions())
-}
-
-// router
-type router struct {
-	config *config.WorkflowConfig
-}
-
-func newRouter(config *config.WorkflowConfig) core.Router {
-	return &router{config: config}
-}
-
-// router interface
-func (r *router) Route(appID string) core.Route {
-	return newRoute(appID, r.config)
-}
-
-func (r *router) Clean() {
-	r.config = nil
-}
-
-// route interface
-type route struct {
-	data sync.Map
-}
-
-func newRoute(appID string, config *config.WorkflowConfig) *route {
-	r := route{
-		data: sync.Map{},
-	}
-	for i, app := range config.Functions {
-		r.Add(i, app.Name)
-	}
-
-	return &r
-}
-
-func (r *route) Add(index int, name string) {
-	logger.Debugf("%sroute add: %s", zipperLogPrefix, name)
-	r.data.Store(index, name)
-}
-
-func (r *route) Exists(name string) bool {
-	var ok bool
-	r.data.Range(func(key interface{}, val interface{}) bool {
-		if val.(string) == name {
-			ok = true
-			return false
-		}
-		return true
-	})
-
-	return ok
-}
-
-func (r *route) Next(current string) (string, bool) {
-	var idx int
-	r.data.Range(func(key interface{}, val interface{}) bool {
-		if val.(string) == current {
-			idx = key.(int) + 1
-			return false
-		}
-		return true
-	})
-	to, ok := r.data.Load(idx)
-	if ok {
-		return to.(string), true
-	}
-
-	return "", false
 }
